ECMP-Configure(Controller/Service/Tests/Auth) feature/add-Entities-and-Endpoints -> Configured ControllerAdvice, Configured DTOs
