+++ ecmp-dev/src/main/java/com/api/ecmpdev/dtos/mappers/ResponseUserMapper.java
@@ -2,11 +2,9 @@ package com.api.ecmpdev.dtos.mappers;
 
 import com.api.ecmpdev.dtos.ResponseUser;
 import com.api.ecmpdev.models.User;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 @Service
 public class ResponseUserMapper implements Function<User, ResponseUser> {
@@ -17,10 +15,7 @@ public class ResponseUserMapper implements Function<User, ResponseUser> {
                 user.getName(),
                 user.getFirstname(),
                 user.getEmail(),
-                user.getAuthorities()
-                        .stream()
-                        .map(GrantedAuthority::getAuthority)
-                        .collect(Collectors.toList())
+                user.getRole()
         );
     }
 }
