+++ ecmp-dev/src/main/java/com/api/ecmpdev/services/UserService.java
@@ -3,18 +3,17 @@ package com.api.ecmpdev.services;
 import com.api.ecmpdev.configs.exceptions.UserEmailNotFoundException;
 import com.api.ecmpdev.configs.exceptions.UserIdNotFoundException;
 import com.api.ecmpdev.dtos.RequestChangeEmail;
-import com.api.ecmpdev.dtos.auth.RequestAuthEmail;
-import com.api.ecmpdev.dtos.auth.RequestAuthId;
 import com.api.ecmpdev.dtos.RequestChangePassword;
 import com.api.ecmpdev.dtos.RequestCreateUser;
 import com.api.ecmpdev.dtos.ResponseUser;
+import com.api.ecmpdev.dtos.auth.RequestAuthEmail;
+import com.api.ecmpdev.dtos.auth.RequestAuthId;
 import com.api.ecmpdev.dtos.mappers.ResponseUserMapper;
 import com.api.ecmpdev.models.User;
 import com.api.ecmpdev.repositories.TokenRepository;
 import com.api.ecmpdev.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
